Allow gen_expr to retry with a different kind
Summary:
When the kind of expression to be generated is `ExprKind::Var`, and there is no variable of the right type available in the environment, we are stuck.
Previously we would emit some dummy, often ill-typed literal in that case. With this diff, we instead retry with a different kind.

Reviewed By: TD5

Differential Revision: D45049419

fbshipit-source-id: b964dc777d3508e9c08d83f98e1052aa7d8b4b20