Better tracking of types
Summary:
Programs that are trivially ill_typed (e.g. `true * []`) tend to abort quickly, reducing how much of the VM they test. Additionally, if the type error is sufficiently obvious it can lead erlc to cut short its analyses.
For these reasons, erlfuzz tries to keep track of a rough approximation of the types of things, to have a decent probability of generating large well-typed fragments.

In this patch, I improved this functionality in various ways:
- keeping track of the types of expressions
- keeping track of the return types of functions, and picking functions to call based on it
- keeping track of the types of arguments of functions for each clause
- etc..

This refactoring will also make it much easier in a later patch to teach erlfuzz about a broader set of bifs.

Reviewed By: TD5

Differential Revision: D45277925

fbshipit-source-id: d491cd57ccb88e7eea25e49cbe6dfb4629c369cc