Generate calls to most functions in the erlang module
Summary:
We were only generating calls to a few kinds of bifs: tests (e.g. `is_integer`) and nullary functions.
Thanks to support for better types, it becomes reasonable to generate calls to all kinds of other bifs.
This is useful mostly for increasing coverage of the VM, but it may also find things in erlc since it has special optimisations for functions like `element` and `setelement`.

Reviewed By: TD5

Differential Revision: D45277924

fbshipit-source-id: 500dd428a6de36c69b00e5e8b89b5d6330eeb133