Mark erlang:external_size as non-deterministic
Summary:
The following testcase:
```
f(_V1) ->
    erlang:external_size(fun(
        <<18446744073709551615:(ok)/integer-unsigned-big, _:(_V1)/integer-little>>
    ) ->
        ok
    end).
wrapper0() ->
    io:write(f(ok)).
```
prints either 81 or 85 depending on which compiler optimisations are enabled. I don't think this is a compiler bug, as external_size only promises to be an upper-bound, not a precise number. And while `term_to_binary` has a `deterministic` option, that is not the case for `erlang:external_size`.

Reviewed By: michalmuskala

Differential Revision: D45779633

fbshipit-source-id: 946dc08073ce1ea0435a6d44b433eef68356f4fe