Reduce the fraction of overloaded functions
Summary:
erlfuzz tries to generate specs for some of the functions it generates. Unfortunately, its understanding of types is too coarse to give interesting specs to overloaded functions, making it rare for specs to be non-trivial.
This patch just reduces the proportion of overloaded functions from 80% to 37.5% to mitigate this problem.

Reviewed By: michalmuskala

Differential Revision: D50359578

fbshipit-source-id: 569dcf08e1077fd13563763081495176063a8c21