Add record creation
Summary:
Note that it is allowed to do record creation in guards, but only if all fields are initialized with an expression that is valid in guards. So if there is a default initializer, we have to remember whether it was generated in a way that guarantees compatibility with guards.
Another part of this diff that may be surprising is that I decided to add a RecordType, rather than just using Tuple. This allows using the names of records in type specs (could unearth issues in dialyzer/eqwalizer treatments of records).

Reviewed By: TD5

Differential Revision: D54201603

fbshipit-source-id: 0c16feae99b472f0d29bd0eb4131dde3300446c9