Add record access
Summary:
We already generated record creation, update, and indexing, but I had forgotten the most basic operation: reading a given field!

This diff fixes this oversight.

There is one small limitation: erlfuzz avoids generating this operation in guards. The reason is that record expansion is different in guards and outside of them, and it blows up nested field accesses in guards by so much that erlc tends to timeout.

Reviewed By: michalmuskala

Differential Revision: D72043916

fbshipit-source-id: b0f0b049b7a496ddc8cf2b8deb9787694b6774f9